edit UpdatePID()+

+++ src/procedure.cpp
@@ -7,7 +7,7 @@ void PID_Init(PIDController *pid, uint16_t Kp, uint16_t Ki) {
     pid->Ki = (float)Ki/10000;
 }
 
-uint8_t UpdatePID(uint8_t cn){
+int16_t UpdatePID(uint8_t cn){
   int16_t error, max = 255, min = -127;
   float output;
   if(settings.sp_structs[0].mode == 4 && cn == 1){  // 4-импульсный режим для канала №2
@@ -28,7 +28,7 @@ uint8_t UpdatePID(uint8_t cn){
   else if (pid[cn].pPart <= min) pid[cn].iPart = 0; // Сброс интеграла
 
   error = output;
-  return (uint8_t)error;
+  return error;
 }
 //------------- симистричный таймер -------------------
 void rotate_trays(void){
