+++ src/procedure.h
@@ -12,7 +12,7 @@ extern uint8_t seconds;
 
 void initMyConfig();
 void PID_Init(PIDController *pid, uint16_t Kp, uint16_t Ki);
-uint8_t UpdatePID(uint8_t cn);
+int16_t UpdatePID(uint8_t cn);
 uint16_t lampUpdate(uint16_t xpos, uint16_t ypos);
 void rotate_trays(void);
 bool check_freeze(uint8_t i);
